pellucid-labeling-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.7
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
spacy==3.7.2
cryptography==41.0.7
numpy==1.24.3
scikit-learn==1.3.0
requests==2.31.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
motor==3.3.2
sse-starlette==1.8.2
python-jose[cryptography]==3.3.0